# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Onboarding MVP API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/mentor/mentees": {
      "get": {
        "tags": [
          "mentor"
        ],
        "summary": "Mentees",
        "operationId": "mentees_mentor_mentees_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MenteeSummary"
                  },
                  "type": "array",
                  "title": "Response Mentees Mentor Mentees Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/mentor/mentees/{user_id}/plan": {
      "get": {
        "tags": [
          "mentor"
        ],
        "summary": "Mentee Plan",
        "operationId": "mentee_plan_mentor_mentees__user_id__plan_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mentor/mentees/{user_id}/tasks/{task_id}/deadline": {
      "patch": {
        "tags": [
          "mentor"
        ],
        "summary": "Update Deadline",
        "operationId": "update_deadline_mentor_mentees__user_id__tasks__task_id__deadline_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskDeadlineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mentor/mentees/{user_id}/tasks": {
      "post": {
        "tags": [
          "mentor"
        ],
        "summary": "Add Task",
        "operationId": "add_task_mentor_mentees__user_id__tasks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MentorTaskCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mentor/mentees/{user_id}/tasks/{task_id}": {
      "delete": {
        "tags": [
          "mentor"
        ],
        "summary": "Delete Task",
        "operationId": "delete_task_mentor_mentees__user_id__tasks__task_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCompleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MenteeSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "full_name": {
            "type": "string",
            "title": "Full Name"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "last_feedback_available": {
            "type": "boolean",
            "title": "Last Feedback Available"
          }
        },
        "type": "object",
        "required": [
          "id",
          "full_name",
          "position",
          "progress_percent",
          "start_date",
          "last_feedback_available"
        ],
        "title": "MenteeSummary"
      },
      "MentorTaskCreate": {
        "properties": {
          "stage_id": {
            "type": "integer",
            "title": "Stage Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          }
        },
        "type": "object",
        "required": [
          "stage_id",
          "title",
          "due_date"
        ],
        "title": "MentorTaskCreate"
      },
      "PlanRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "template_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Template Id"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "progress_percent": {
            "type": "integer",
            "title": "Progress Percent"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRead"
            },
            "type": "array",
            "title": "Stages",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "template_id",
          "start_date",
          "progress_percent"
        ],
        "title": "PlanRead"
      },
      "StageRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "order_index": {
            "type": "integer",
            "title": "Order Index"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskRead"
            },
            "type": "array",
            "title": "Tasks",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "title",
          "order_index"
        ],
        "title": "StageRead"
      },
      "TaskCompleteResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "TaskCompleteResponse"
      },
      "TaskDeadlineUpdate": {
        "properties": {
          "due_date": {
            "type": "string",
            "format": "date",
            "title": "Due Date"
          }
        },
        "type": "object",
        "required": [
          "due_date"
        ],
        "title": "TaskDeadlineUpdate"
      },
      "TaskRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "stage_id": {
            "type": "integer",
            "title": "Stage Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "due_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due Date"
          },
          "is_completed": {
            "type": "boolean",
            "title": "Is Completed"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "is_system_task": {
            "type": "boolean",
            "title": "Is System Task"
          },
          "requires_confirmation": {
            "type": "boolean",
            "title": "Requires Confirmation"
          },
          "added_by_mentor_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Added By Mentor Id"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "stage_id",
          "title",
          "description",
          "due_date",
          "is_completed",
          "completed_at",
          "is_system_task",
          "requires_confirmation",
          "added_by_mentor_id",
          "updated_at"
        ],
        "title": "TaskRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}